arkeovision_backend/requirements.txt
django>=4.2,<5.1
djangorestframework>=3.14
django-cors-headers>=4.3
python-dotenv>=1.0
openai>=1.30
requests>=2.31
